database/requirements.txt
python-dotenv
openai
tiktoken
docling
lancedb
streamlit